fb api implemented
